backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy[asyncio]==2.0.23
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-core==0.1.8
openai>=1.6.1,<2.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
alembic==1.12.1
httpx==0.25.2
requests==2.31.0
email-validator==2.1.0.post1
pandas==2.1.3
beautifulsoup4==4.12.2
matplotlib==3.8.2
psycopg2-binary==2.9.9
redis==5.0.1
asyncpg==0.29.0 